requirements.txt
grpcio==1.10.1
grpcio-tools==1.10.1
protobuf==3.5.2.post1
psycopg2==2.7.4
psycopg2-binary==2.7.4
six==1.11.0
